Add auto-initialization for git submodules
A condition has been added to check if git submodules have been initialized before the build. If not, they are now automatically initialized and a ".gitsubmoduleinit" file is created for future checks. This has been done to ensure smooth building of the project without manual submodule initialization. Additionally, the ".gitsubmoduleinit" file has been added to the .gitignore file and the redundant submodule checkout script in ".space.kts" has been removed.