[PKGS-1056] "Use readAndWriteAction in MppDependencyModifier"
The adjustment made to the function modifyKotlinModel in MppDependencyModifier.kt includes implementing both read and write actions together (readAndWriteAction). Previously, only readAction was used, and an additional write action was needed to apply the changes. This change ensures that the build model can be modified and the changes saved immediately, reducing possibilities of inconsistency or data loss due to unexpected interruptions.