Update publishing channels and Gradle script command
This commit changes the publishing channel in the build.gradle.kts from 'EAP' to 'INTERNAL-EAP'. The purpose of this change is to target a more specific, internal audience for early testing.

In .space.kts, the shellScript content for running Gradle has been updated to './gradlew publish :plugin:publishPlugin'. This addition specifically includes the :plugin:publishPlugin to make sure that the plugin is also published along with the main project.