Update plugin settings and description
The build script is updated to allow the plugin token to be read from project properties or environment variables. This allows for more flexible configurations in different environments. Additionally, the plugin ID is changed to better represent its source. The description is expanded to provide more information about the plugin's functionality and supported projects.