"Refactor code to enhance readability and maintainability"
In this commit, we have refactored several points in the code to more clearly express its intent and improve maintainability. This mostly consists of changing the usage of `val` to `const val` for constant variables, removing TODO items that were redundant, and streamlining some lines of code. Along with these, we also introduced the extension of `WithVariant` in `PackageSearchDeclaredPackage` to include a `variantName` for each declared dependency, enhancing the expressiveness of our data objects. This ensures that our codebase remains manageable and expressive.