Refactor plugin code structure for better readability and performance
This commit includes several changes in the IntelliJ Package Search Plugin code structure, with the goal to enhance code readability and improve performance.

The main changes include refactoring of `PackageGroup.kt` and `SearchData.kt` to use selected modules directly rather than handling them in search data. It also included a fix in `PackageItemRow.kt` for a display issue where the action content was potentially cut off.

Adjustments were made for the coroutine flow in `GradleModuleTransformer.kt` and `GradleUtils.kt` to improve the flow of data. Also, added a new flow named `fileOpenedFlow` in `Utils.kt` to get a flow of files when they are opened in the IDE.

Changes in `PackageSearchPackagePanel.kt` were made to remove unnecessary warning logs that were originally logged during the package search process. The naming of `ResultsColumn.kt` was changed to `PackageSearchPackageList.kt` for better class enunciation.

Lastly, a small but meaningful fix in `PackageSearchKotlinMultiplatformVariant.kt` was made to stem off the return of attributes when it's empty.

Overall, these changes make the code cleaner, more understandable and should provide a small performance boost in runtime.