Update IntelliJ version and implement Disposable for caches
Updated the default IntelliJ version from "232-SNAPSHOT" to "2023.2.2" in the PackageSearchExtension. This change was made to keep the project up-to-date with the latest IntelliJ version.

Also, updated the PackageSearchProjectCachesService to implement the Disposable interface, which allows the caches to clean up and release resources when no longer needed. This change is expected to minimize our service's memory footprint by ensuring appropriate resource management.