Implement improvements to Package Search Plugin
This major commit includes improvements to the JetBrains Package Search Plugin and streamlines badges for Kotlin Multiplatform Projects (KMP).

We are making changes throughout the Package Search Plugin's code base. Here's a brief outline of the changes:

1. Added an import for the "flatten" method in PackageGroup.kt and improved how “badges” are constructed from compatible variants.
2. Included "logback" to the list of packages in PackageSearchExtension.kt.
3. Improved dependency management by incorporating a "gradleModel" in PackageSearchGradleDependencyManager.kt.
4. Augmented the handling of Unified Dependency in PackageSearchGradleDependencyManager.kt and included a "toGradleDependencyString" function.
5. Modified the url in tests.http.
6. Integrated "model" into PackageSearchKotlinMultiplatformDependencyManager and updated the build Attributes function in KotlinMultiplatformModuleProvider.kt.
7. Additional updates are made to other files for enhancing the functionality and resolving minor issues.

The primary motive behind these changes is to improve the visualisation of the KMP metadata in the Package Search UI in IntelliJ IDEA, and to improve how badges are constructed and managed.