Remove unnecessary code and adjust plugin compatibility range
This commit removes some unnecessary parts of the code, specifically in `.space.kts` and `build.gradle.kts`. Commented out "toolbox-enterprise-token" in `.space.kts` and "configurations.api" block in `build.gradle.kts` are removed as they are not needed anymore. Meanwhile, as for ensuring correct plugin compatibility, "untilBuild" and "sinceBuild" attributes were added under "patchPluginXml" in `build.gradle.kts`, signaling that the plugin is compatible with the range from version 232.2.* to 233.*. Lastly, `group` under "publishShadowPlugin" is updated to "publishing" to reflect its correct categorization.