Set dynamic plugin version in build script
Added a line in the 'patchPluginXml' section of the build.gradle.kts file for the plugin module to dynamically assign the version of the plugin using environment variable `PLUGIN_VERSION`. If this variable is not present, the project's version is used as a fallback. This allows for more flexible versioning and deployment procedures.