Refactor MoreActionPopupContent and PackageItemRow for better UI interaction
This commit updates the UI handling in MoreActionPopupContent and PackageItemRow by adding popupOpenStatus to reset the popup status after an action. This improves the user's experience of the navigation and interaction with the popup. Also, introduced a clickable attribute to elements in PackageItemRow replacing the pointer input providing seamless UI behavior. Lastly, updated Box in PackageItemRow with svg icon 'more.svg'.