Refactor build and publish scripts for new jewel version
We updated the build and publish related files and scripts to accommodate a newer version of jewel; '.space.kts' and 'ConfigurePublishPlugin.kt' has been refactored to modify the 'Publish snapshots' job and include shadowJar task in MavenPublication respectively. 'settings.gradle.kts' and 'packagesearch.versions.toml' have their dependencies updated with the new version of jewel. The newer version of jewel also required an addition to the gradle configuration in the project, reflected by changes in the 'ConfigurePublishPlugin.kt' script file. This change will streamline the build and publish process with the updated version of the jewel library.