Updated Gradle cache, adjusted delay & debounce, improved packages and edited UI style
In this commit, several changes were made across multiple files to improve build performance, package management, and user interface styling. The build caching system was modified in .space.kts to include '**/.gradle' and '**/build', aiming to enhance build speed. The `delay` and `debounce` times in PackageSearchProjectService.kt were increased for better performance and smoother user experience. In packagesearch.versions.toml, the `composeDesktop` version was updated to "1.5.1". Also, the call to `pxToDp()` was replaced with `dp` directly in the SearchRow.kt file in the user interface to streamline and simplify the code. Lastly, the jewel-int-ui-standalone project path was corrected for a better organization in settings.gradle.kts.