Update plugin compatibility and publication settings
The compatibility value for 'sinceBuild' in the plugin's XML patch was changed to a fixed version '232' instead of all sub-versions of '232'. This modification was done to specify a more accurate compatibility standard.

In the 'ConfigurePublishPlugin.kt' file, the publication setting was changed from 'kotlin' to 'java'. This change ensures that the publication artifact is correctly created for java components.