Refine Gradle caching in space.kts file
The cache configuration in the space.kts file has been updated to have more specific cache directories for the builds. Instead of using a single cache for all Gradle builds and another for all builds, now different cache sections have been specified for individual parts of the application such as 'root', 'jewel', 'buildSrc' and 'api-models'. This change aims to enhance build speeds by separating the caching of different parts of the project.