Inject SoftwareComponentFactory and refactor PublishPlugin configuration
Added SoftwareComponentFactory as a dependency in PackageSearchPlugin and used it within the configuration of the 'PublishPlugin'. This change allows for more control over the publication configuration, specifically it allows the addition of a shadowJar to the Maven publication. Furthermore, it simplifies the configuration by removing redundant evaluations and moving the configuration logic into a function.