Update build.gradle.kts to use "Release" channel
Changed the Gradle script to use the "Release" channel instead of "INTERNAL-EAP". This is in line with our move towards making our plugin available for wider audience and not just for internal testing. The change should not affect the functionality but will ensure that more stable versions of the plugin are being used.