Update plugin config and refactor deployment process
Made the plugin usage more convenient by removing the necessity of a restart every time the plugin is loaded. This is achieved by updating the 'plugin.xml' file. Furthermore, added a null protection mechanism in the corresponding '.space.kts' file to ensure the deployment process doesn't fail when the 'build-scan-url.txt' file is missing or empty. Both changes aim to improve the user experience and the deployment efficiency.