Update automation versioning for deployments
The automation versioning in '.space.kts' script was modified to use 'api.executionNumber()' instead of parameter 'version'. This change was made to handle the incremental versioning of automation deployments more robustly, and ensures that each deployment gets an unique version number based on the execution number rather sticking to a manual entry.