Update build-scan-url file location search
Modified the search method for the "build-scan-url.txt" file in the '.space.kts' script to walk the directory top down to a maximum depth of 2, instead of checking if it exists in the root directory only. This change ensures that the files located in sub-directories are also identifiable, making the script more reliable.