Added a state to control enabled/visibility of the InfoBox component
This commit introduces the 'LocalInfoBoxPanelEnabled' state in order to manage the visibility of the InfoBox in the 'PackageSearchPackageList' component. Now, the InfoBox panel will not be shown to users when it's not enabled from 'PackageSearchToolWindowFactory'. This change was implemented to improve UI interactions and to give better control over what is displayed in the window, enhancing the usability.