"Improve error handling and dependency search functionality"
This commit introduces several changes to increase error handling robustness and improve the functionality of dependency searches.

A critical error logging feature was added to the 'PackageSearchApiClientType' file to handle issues while retrieving package information from Sonatype. Additionally, a systematic refresh was introduced in 'PackageSearchApplicationCachesService' to refresh dependencies whenever an engine failure happens.

The tree builder in 'Utils.kt' was optimized by eliminating redundant 'nodesIds' building.

To enhance user experience, two message lines were introduced in 'PackageSearchPackageList' in place of one to handle a case when no supported dependencies were found, making it more informative.

Due to failures of retrieving packages, retry feature on HTTP requests was added in 'SonatypeApiClient' accompanied by an error handling method in 'PackageSearchSonatypeApiClient'. This helps in cases of network or server side failures, subsequently improving the stability in retrieving packages data.

Lastly, an additional error handling was added in 'PackageSearchToolWindow' for the case when the tree builder returns an empty tree.