Add automated job for checking IJ snapshot errors
This commit introduces a new automated job, "Check IJ snapshot errors". This job is scheduled to run daily at 8 AM through a cron schedule. The job executes a shell script to build the shadow plugin. The commit also modifies the "intellijVersion" to drop the "EAP-" prefix, reflects the removal of specific build bounds in the 'patchPluginXml' configuration and slightly adjusts the way configuration observation is handled in 'PackageSearchProjectService'. These changes are to ensure enhanced development workflow and better error handling.