Refactor Gradle task execution in .space.kts for deployment
This update refactors the execution of Gradle tasks in the .space.kts file to better handle the deployment process. In particular, the Gradle tasks run now serially through the `api.runCatchingGradlew` method. Furthermore, a change was made in the conditional statement to accurately check the success of the task execution. Also, the addition of the `GradleCommandBuilder` class provides a cleaner and more efficient way to construct the Gradle task commands. This change fixes an earlier issue related to incorrect method usage and facilitates error-free deployment. Especially, this change is essential in ensuring our plugin versioning is correctly handled during deployment.