Refactor BaseGradleModuleProvider to AbstractGradleModuleProvider
The BaseGradleModuleProvider has been renamed to AbstractGradleModuleProvider. This includes changes in all classes where BaseGradleModuleProvider was extended. This was done to better reflect its use as a superclass and align with typical naming conventions in object-oriented programming where 'abstract' is used to denote classes meant to be subclassed. BaseGradleModuleProvider has been kept as a typealias and deprecated to maintain backward compatibility.