Refactor API call handling and error flow control
This commit removes the use of runCatching for API calls in PackageSearchApiPackageCache and instead handles potential failures further upstream. Additionally, introduced in the PackageSearchProjectService is a restart mechanism with a retry limit that prevents continuous restarts and avoids potential infinite loops. The PackageSearchGradleModelBuilder class also had an unused import that was removed.