Update keys selection logic in PackageSearchModulesTree
The method of determining and storing selected keys in 'PackageSearchModulesTree.kt' file has been revised. Particularly, the selected keys are now directly assigned the intersection result without first converting to a list. Additionally, the first taken result from 'walkBreadthFirst()' function is now converted to a set instead of a list.