Refactor package search splitter for improved handling
The refactoring of the "packageSearchSplitter" method in the Components.kt file has improved the handling of this functionality. By introducing a new Divider component with a Vertical Orientation, the need for a hidden state check was eliminated, simplifying the code and significantly improving readability and maintainability.

(cherry picked from commit b0281ec4a8978e870382e4d2789196e38d76a3bc)