241 - Improved services lazyness and reduced HTTP client retry attempts. (#73)
* Improved services lazyness and reduced HTTP client retry attempts.

* Remove unused gradleIdentityPathOrNull utility

This commit removes the unused 'gradleIdentityPathOrNull' utility from the Utils.kt file. This method was previously used to retrieve Gradle module identity path, but is no longer needed. The appearance and readability of the code have significantly improved as a result.

* Refactor PackageSearchApiPackageCache handling

The code for handling the network results in the PackageSearchApiPackageCache class has been refactored. Instead of performing a removal of old entries followed by an insertion of new entries, the code now performs an update operation for each new entry. This makes the handling of network results more efficient and concise.

* Update trigger in snapshot publish workflow

The trigger in the GitHub Actions workflow `.github/workflows/publish-snapshot.yml` has been updated to only initiate upon pushes to the `releases/241` branch. This ensures snapshots are only published for relevant release branches.