Fix Gradle command in test workflow (#150)
* Update Gradle command in test workflow

The command used to run Gradle in the testing GitHub Action has been updated. Previously, it was using inputs for the task and options separately. Now, it uses a single input for the entire command, which increases flexibility.

* Update validation tests in PackageSearch plugin

Revised the validation tests in PackageSearch plugin by using Kotlin's standard testing functions, improving readability. The old 'assert' methods used have been replaced by 'assertEquals' and 'assertTrue' where relevant. This also ensures test failure messages are more descriptive and helpful.

* Refactor GitHub actions to simplify test workflow

The previous GitHub workflow configuration file was deleted and replaced with a simpler and direct test.yml file. Specifically, this refactor consolidates the setup and execution tasks for Gradle, Maven, and KMP within the test workflow instead of running them as separate actions. Additionally, the test results are directly uploaded as part

* Refactor GitHub actions to simplify test workflow

The previous GitHub workflow configuration file was deleted and replaced with a simpler and direct test.yml file. Specifically, this refactor consolidates the setup and execution tasks for Gradle, Maven, and KMP within the test workflow instead of running them as separate actions. Additionally, the test results are directly uploaded as part

* Update artifact naming in GitHub Actions workflow

This commit modifies the name structure of the test result artifacts in the GitHub Actions workflow. These changes add clarity by including the build tool (Gradle, Maven, or KMP) in the artifact name, which