Add GitHub workflow task for version-catalog test
This update includes a new GitHub workflow in test.yml to run 'version-catalog' tests across multiple OS and Gradle versions. Additionally, this commit involves renaming MavenProjectTest to VersionCatalogTest and adding new test resources for Gradle catalog projects.The name of the workflow specified in the .github/workflows/test.yml file has been updated. Previously, this was tagged as "Gradle", but it has been changed to "Version Catalog" to give a more