Add GitHub Workflow task for version-catalog test (#156)
* Update test workflow and add new config file

Expanded the test workflow in GitHub Actions to include the specific 'fabrizio.scarponi/extend-test-scenario' branch. In addition, a new configuration file named 'spring-groovy.json' was created for

* Add GitHub workflow task for version-catalog test

This update includes a new GitHub workflow in test.yml to run 'version-catalog' tests across multiple OS and Gradle versions. Additionally, this commit involves renaming MavenProjectTest to VersionCatalogTest and adding new test resources for Gradle catalog projects.The name of the workflow specified in the .github/workflows/test.yml file has been updated. Previously, this was tagged as "Gradle", but it has been changed to "Version Catalog" to give a more

* Add GitHub workflow task for version-catalog test

This update includes a new GitHub workflow in test.yml to run 'version-catalog' tests across multiple OS and Gradle versions. Additionally, this commit involves renaming MavenProjectTest to VersionCatalogTest and adding new test resources for Gradle catalog projects.The name of the workflow specified in the .github/workflows/test.yml file has been updated. Previously, this was the workflow purpose.

* Refactor test class names and related workflow settings

Rename `MavenProjectsTest` to `CatalogProjectTest`, adjust its resource path accordingly, and reflect this change in the Github workflow file. Rename `VersionCatalogTest` to `MavenProjectTest` and update its resource path, achieving a better alignment with the functionality being tested.

* Allow test workflow to continue on error

A setting has been added to the GitHub test workflow configuration to allow it to continue running even after encountering errors. This will ensure that a single error doesn't halt the entire testing process for the version-catalog.

* Update naming for GitHub actions and artifacts

The names for the GitHub actions and artifacts related to the Kotlin Multiplatform (KMP) have been updated to include "Gradle

* Reduce test pool for Catalog tests