233 - Discontinue Test (#152)
* Fix Gradle command in test workflow (#150)

* Update Gradle command in test workflow

The command used to run Gradle in the testing GitHub Action has been updated. Previously, it was using inputs for the task and options separately. Now, it uses a single input for the entire command, which increases flexibility.

* Update validation tests in PackageSearch plugin

Revised the validation tests in PackageSearch plugin by using Kotlin's standard testing functions, improving readability. The old 'assert' methods used have been replaced by 'assertEquals' and 'assertTrue' where relevant. This also ensures test failure messages are more descriptive and helpful.

* Refactor GitHub actions to simplify test workflow

The previous GitHub workflow configuration file was deleted and replaced with a simpler and direct test.yml file. Specifically, this refactor consolidates the setup and execution tasks for Gradle, Maven, and KMP within the test workflow instead of running them as separate actions. Additionally, the test results are directly uploaded as part

* Refactor GitHub actions to simplify test workflow

The previous GitHub workflow configuration file was deleted and replaced with a simpler and direct test.yml file. Specifically, this refactor consolidates the setup and execution tasks for Gradle, Maven, and KMP within the test workflow instead of running them as separate actions. Additionally, the test results are directly uploaded as part

* Update artifact naming in GitHub Actions workflow

This commit modifies the name structure of the test result artifacts in the GitHub Actions workflow. These changes add clarity by including the build tool (Gradle, Maven, or KMP) in the artifact name, which

(cherry picked from commit 84e01e15fad231714e167b9cb773cfaba9e684d3)

* Update test case to use newer release

Deleted the MavenProjectsTest.kt file and updated the Utils.kt file to use a new release "233.3.6" instead of "233.3.5". The removal of the test case was necessary as it is no longer relevant, and the update in the Utils file ensures tests are run in the latest environment.

* Downgrade test case project version

The project version in the Utils test case has been downgraded from "233.3.6" to "233.3.5". This change was needed to ensure the correct functioning of the API module after a recent update.

* Disable GitHub workflow test for 233

The file .github/workflows/test.yml has been renamed to .github/workflows/test.yml.disabled, effectively disabling the associated GitHub workflow. This change is likely made to temporarily halt automatic testing until necessary updates or changes are made.

---------

Co-authored-by: Lamberto Basti <basti.lamberto@gmail.com>