Add test context to project service tests and update dependencies
Extended the project service tests (KMPGradleProject, MavenProject, GradleProject) to include a private test context. This alteration provides a consistent way to interact with the IDE context within the tests. Additionally, dependencies have been updated, including a version change for the 'ideStarter' and the removal of unnecessary maven credentials from the build script.